Update stock_exchange.sql
Completed the stored procedure for executing the sale.
Changed all the datetime to datetime2
Completed the procedure to delete an order
Changed the name for finished_orders to executed orders.
Moved the stored procedure exe_buy_sell before the new_order so that the new_oeder procedure can understand the other exists.