Buyers and sellers tables were removed, Table creation and procedures were separated
The original file was split in two:
stock_exchange_table_creation that  creates all the necessary tables, and
stock_exchange_execution that has the procedures to execute and/or stores the buy/sell orders.

The tables are also limited to just orders, deleted_orders, and executed_orders. The buyers and sellers tables from previous version are not used any more. This makes the code more streamlined and reduces the number of insert and delete operations, and the number of if statements necessary for decision making between which tables to delete from.